# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/articles": {
      "get": {
        "operationId": "ArticlesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List articles with search & pagination",
        "tags": [
          "Catalog - Articles"
        ]
      },
      "post": {
        "operationId": "ArticlesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create article",
        "tags": [
          "Catalog - Articles"
        ]
      }
    },
    "/api/v1/articles/{id}": {
      "get": {
        "operationId": "ArticlesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get article by ID",
        "tags": [
          "Catalog - Articles"
        ]
      },
      "patch": {
        "operationId": "ArticlesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update article (field-level RBAC)",
        "tags": [
          "Catalog - Articles"
        ]
      }
    },
    "/api/v1/articles/{id}/bom": {
      "get": {
        "operationId": "ArticlesController_getBom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Состав изделия (BOM): из чего оно собирается",
        "tags": [
          "Catalog - Articles"
        ]
      },
      "post": {
        "operationId": "ArticlesController_addBomItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Добавить позицию состава (расход материала на единицу)",
        "tags": [
          "Catalog - Articles"
        ]
      },
      "put": {
        "operationId": "ArticlesController_replaceBom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Заменить состав целиком («собрать ГП»: полный список расхода)",
        "tags": [
          "Catalog - Articles"
        ]
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "operationId": "MaterialsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List materials with search & pagination",
        "tags": [
          "Catalog - Materials"
        ]
      },
      "post": {
        "operationId": "MaterialsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create material",
        "tags": [
          "Catalog - Materials"
        ]
      }
    },
    "/api/v1/materials/{id}": {
      "get": {
        "operationId": "MaterialsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get material by ID",
        "tags": [
          "Catalog - Materials"
        ]
      },
      "patch": {
        "operationId": "MaterialsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update material (field-level RBAC)",
        "tags": [
          "Catalog - Materials"
        ]
      }
    },
    "/api/v1/customers": {
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "Catalog - Customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer",
        "tags": [
          "Catalog - Customers"
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get customer by ID",
        "tags": [
          "Catalog - Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer",
        "tags": [
          "Catalog - Customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "ЦМК АВРОРА — ERP REST API",
    "description": "REST API v1 ЦМК АВРОРА — планирование производства, заказы, склад, финансы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}